Updated the readme with important notes about the python-gnupg module and the gnupg fork.
Also the licensing factor (fucking idiot move of using the GPL for
something intended to be used as a library, it's PyQt all over again).
Still, this code can be used without either module, all it takes is
running the commands directly and a little fiddling for OS detection,
so if I absolutely have to that's what I'll do.  Alternatively, using
a symmetric cipher with openssl is viable.