Removed python-gnupg and all dependence on GPG or any wrapper library.
Replaced with a slightly modified version of simplecrypt and PyCrypto.
The simplecrypt modification increases the number of rounds used from
10,000 to 250,000.

Thus neatly avoiding the impending disaster that is the python-gnupg
vs. gnupg thing.

Note to self: check the feasibility of blowing both packages out of
the water with a full port of gpme when gpg 2.1 is ready.