Updated Tor config to use SOCKS5 instead of old style HTTP/S and privoxy.
Now asks if user is using Tor Browser Bundle (if yes then connection
is set to port 9150, if no then connection is set to port 9050).
Still requires a local instance of Tor to run.  May change to
automatic configuration in the future (by sending a HTTPS request to
check.torproject or some other .onion address and checking the status
code through requests.