Moved the beginning work on new (argparse) calling method to its own directory.
For some reason trying to move this into a separate branch didn't work
so screw it, I'll build on it in a separate space to the loose
collection of scripts.  I may end up making it far more modular that
way anyway.  It will make it easier to add newer API functions or
other plugins later too.